delete관련 로직 틀 완성

+++ src/main/java/com/semi/controller/DeleteController.java
@@ -29,20 +29,39 @@ public class DeleteController extends HttpServlet {
 			throws ServletException, IOException {
 			// 게시글 번호
 			String id = request.getParameter("id");
+			int id2 = Integer.parseInt(id);
 		
 			WriterDTO dto = new WriterDTO();
-			WriterService service = new WriterService();
-			List<WriterDTO> deleteWriter = service.select(id);
+			dto.setId(id2);
 			
+			WriterDAO dao = new WriterDAO();
+			WriterService service = new WriterService();
+
+			if(dao.delete(dto)) {
+				// 삭제 완료
+				System.out.println("오류1");
 				if(service.delete(dto)) {
-					// 삭제 완료 페이지 띄움
-					String view = "/WEB-INF/jsp/delete/delete.jsp";
+					// 삭제 완료 후 커밋까지 완료
+					System.out.println("오류2");
+					String view ="/WEB-INF/jsp/delete/delete.jsp";
 					RequestDispatcher rd = request.getRequestDispatcher(view);
 					rd.forward(request, response);
 				} else {
-					// 삭제 실패
-					System.out.println("오류");
+					// 삭제 완료 뒤 커밋에서 문제
+					System.out.println("오오류3");
+					request.setAttribute("error", "해당 게시글 삭제 여부를 확인하세요.");
+					String view ="/WEB-INF/jsp/delete/delete.jsp";
+					RequestDispatcher rd = request.getRequestDispatcher(view);
+					rd.forward(request, response);
 				}
+			} else {
+				// 삭제 실패	
+				System.out.println("오류4");
+				request.setAttribute("fail", "게시글 삭제가 실패했습니다.");
+				String view ="/WEB-INF/jsp/delete/delete.jsp";
+				RequestDispatcher rd = request.getRequestDispatcher(view);
+				rd.forward(request, response);
 			}
+		}
 	}
